+++ src/dataprocessor/property/gt_intproperty.h
@@ -15,6 +15,7 @@
 #include "gt_datamodel_exports.h"
 
 #include "gt_property.h"
+#include "gt_boundaries.h"
 
 class GT_DATAMODEL_EXPORT GtIntProperty : public GtProperty<int>
 {
@@ -54,6 +55,7 @@ public:
                   const QString& brief,
                   int value);
 
+    [[deprecated("Use function with GtIntProperty::Boundaries definition instead")]]
     GtIntProperty(const QString& ident,
                   const QString& name,
                   const QString& brief,
@@ -61,11 +63,18 @@ public:
                   int highSideBoundary,
                   int value = 0);
 
+    [[deprecated("Use function with GtIntProperty::Boundaries definition instead")]]
     GtIntProperty(const QString& ident,
                   const QString& name,
                   const QString& brief,
                   BoundType boundType,
                   int boundary,
+                  int value = 0);
+
+    GtIntProperty(const QString& ident,
+                  const QString& name,
+                  const QString& brief,
+                  gt::Boundaries<int> bounds,
                   int value = 0);
 
     /**
@@ -212,6 +221,7 @@ gt::PropertyFactoryFunction makeIntProperty(int value);
  * @param lowSideBoundary Lower side boundary
  * @param highSideBoundary High side boundary
  */
+[[deprecated("Use function with GtIntProperty::Bound bound definition instead")]]
 GT_DATAMODEL_EXPORT
 gt::PropertyFactoryFunction makeIntProperty(int lowSideBoundary,
                                             int highSideBoundary,
@@ -224,9 +234,14 @@ gt::PropertyFactoryFunction makeIntProperty(int lowSideBoundary,
  * @param boundaryType Boundary type
  * @param boundary Boundary
  */
+[[deprecated("Use function with GtIntProperty::Boundaries definition instead")]]
 GT_DATAMODEL_EXPORT
 gt::PropertyFactoryFunction makeIntProperty(GtIntProperty::BoundType boundaryType,
                                             int boundary,
+                                            int value = 0);
+
+GT_DATAMODEL_EXPORT
+gt::PropertyFactoryFunction makeIntProperty(gt::Boundaries<int> bounds,
                                             int value = 0);
 
 } // namespace gt